Pin FluentAssertions version (#79)
* Bump dependencies

Enable built-in .NET analyzers on .NET Framework. see
https://github.com/dotnet/roslyn-analyzers?tab=readme-ov-file#microsoftcodeanalysisnetanalyzers

* Bump build TargetFramework

* Pin FluentAssertions version

This commit pins FluentAssertions version to
the last version released under an Apache 2.0 license.
The project has been relicensed from version 8.0.0 onwards
to a paid license for commercial use. See

https://github.com/fluentassertions/fluentassertions/pull/2943

I am not a lawyer, but the safest actions are to

- pin to last Apache 2.0 version
or
- remove the use of the library.

This commit does the former. The latter can also be done in
the future.

---------

Co-authored-by: Anush <anushshetty90@gmail.com>